Do not have "Entry" representation in the Store
Calendar entries is difficult to express in store entries.
Loading specific calendar entries is still possible (or rather: will
be), as with icalendar, each entry has a UID. We can load entries with a
specific UID by iterating through all calendar files and search them for
the UID... so this is not a problem.

Either way, Calendar files (.ical) often only contain one entry, so
basically one calendar with one entry... so this is not a problem. If it
becomes a problem, we can redesign this.