Add WithStoreIterator trait
This patch changes the store interface quite a bit:

Before, we had the possibility to create a `StoreIdIteratorWithStore`
by calling `Store::entries()`.

This was nice. The `StoreIdIteratorWithStore` had functions to transform
it into a `StoreCreateIterator` for example.

With this change, we have an abstraction so that _every_ iterator which
holds a reference to the `Store` can be transformed into a
`StoreCreateIterator` (for example). The iterator object has to
implement the `WithStoreIterator` trait for this, which is quite simple.

The `StoreIdIteratorWithStore` iterator type was changed to not provide
the `into_create_iter()` (for example) functions via the type but the
trait,... which changes the interface of `libimagstore`.

This change was made because we ran into the following problem: We
wanted to wrap a `StoreIdIterator` for filtering for a specific
collection in the store ("/calendar", to be specific).
This was possible, of course. But we cannot use this new iterator type
(`CalendarIter`) and call `into_get_iter()` on it.
We would need to implement it directly on the type.
With this change, we get this functionality by implementing the much
less complicated `WithStoreIterator` trait.