Move "backend" to cache
So that accessing the "backend" actually happens through accessing the
"cache".

This way we do not have a "cache" next to the "backend" and the Store
handling access to both, but rather the cache handling the backend access.

This patch is the first step towards a Serialize/Deserialize Cache which
can be read and written to stdin/stdout.